requirements.txt
pillow
django-rest-framework
django-crispy-forms